refactor: 번외, Swagger 패키지 생성 후 위치 변경 (#108)
