+++ src/main/java/com/codenear/butterfly/member/presentation/NicknameController.java
@@ -3,6 +3,7 @@ package com.codenear.butterfly.member.presentation;
 import com.codenear.butterfly.global.dto.ResponseDTO;
 import com.codenear.butterfly.global.util.ResponseUtil;
 import com.codenear.butterfly.member.application.NicknameService;
+import com.codenear.butterfly.member.presentation.swagger.NicknameControllerSwagger;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
